Split our WorldEditorHelper into multiple files
